+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ## [Unreleased]
 ### Security
 ### Added
+- ConsoleCapture test helper in FunFair.Test.Infrastructure.Helpers to redirect Console.Out and Console.Error during tests
 ### Fixed
 ### Changed
 ### Removed